extract ActiveRecordCompat in its own file
